fix: PostgreSQL compatibility for model search scopes
- HelpContent::scopeSearch() - Use LOWER() for case-insensitive search
- LogEntry::scopeSearch() - Use LOWER() for case-insensitive search
- HelpContentRelated - Change relevance_score cast from float to integer
- tests/bootstrap.php - Add opcache_reset() for test reliability
- Update CHANGELOG.md and TESTING.md for v6.9.2

All 645 Model unit tests now passing on PostgreSQL.

+++ app/Models/HelpContentRelated.php
@@ -22,7 +22,7 @@ class HelpContentRelated extends Model
     ];
 
     protected $casts = [
-        'relevance_score' => 'float',
+        'relevance_score' => 'integer',
     ];
 
     /**
